chore(transformer-directives): deconstruct index
